refactor(core)!: shrink GatewayInfo::expiry_delta to u16
A CLTV delta in blocks — the LN protocol itself carries
cltv_expiry_delta as a u16.

+++ picomint-client/src/ln/mod.rs
@@ -45,7 +45,7 @@ use self::send_sm::{SendSMCommon, SendSMState, SendStateMachine, SendStateMachin
 /// Maximum total contract lock, in blocks, the client is willing to accept
 /// from a gateway. Backstop against an abusive gateway tying funds up before
 /// the unilateral refund path opens.
-const EXPIRY_DELTA_LIMIT: u64 = 1000;
+const EXPIRY_DELTA_LIMIT: u16 = 1000;
 
 /// A two hour buffer in case either the client or gateway go offline
 const CONTRACT_CONFIRMATION_BUFFER: u32 = 12;
@@ -225,7 +225,7 @@ async fn send_inner(
         amount,
         fee,
         expiry: consensus_block_count
-            + gateway_info.expiry_delta as u32
+            + u32::from(gateway_info.expiry_delta)
             + CONTRACT_CONFIRMATION_BUFFER,
         claim_pk: gateway_info.module_public_key,
         refund_pk: refund_keypair.x_only_public_key().0,
